docs: 로그인/로그아웃 버튼 header로 이동
