+++ src/app/day/page.tsx
@@ -1,5 +1,3 @@
 export default function DayPage() {
-  return (
-    <div>page</div>
-  );
+  return <div>page</div>;
 }
